+++ src/main/java/controllers/EmployeeController.java
@@ -1,5 +1,7 @@
 package controllers;
 
+import com.jfoenix.controls.JFXTextField;
+import configuration.DataValidation;
 import dao.EmployeeDAO;
 import entity.Employee;
 import javafx.collections.FXCollections;
@@ -20,8 +22,6 @@ import java.io.IOException;
 import java.time.LocalDate;
 
 public class EmployeeController {
-
-
     //--------------TABLE-----------------//
     @FXML
     private TableView<Employee> employeeListTableView;
@@ -37,89 +37,99 @@ public class EmployeeController {
     private TableColumn<Employee, Double> employeeListTableViewSalary;
     //--------------CREATE-----------------//
     @FXML
-    private TextField firstName;
+    private JFXTextField firstName;
     @FXML
-    private TextField lastName;
+    private JFXTextField lastName;
     @FXML
-    private TextField salary;
+    private JFXTextField salary;
     @FXML
-    private TextField email;
+    private JFXTextField email;
     @FXML
     private DatePicker birthDate;
     //--------------UPDATE-----------------//
     @FXML
-    private TextField firstNameUpd;
+    private JFXTextField firstNameUpd;
     @FXML
-    private TextField lastNameUpd;
+    private JFXTextField lastNameUpd;
     @FXML
-    private TextField emailUpd;
+    private JFXTextField emailUpd;
     @FXML
-    private TextField salaryUpd;
+    private JFXTextField salaryUpd;
     @FXML
     private DatePicker birthDateUpd;
-
     //buttons for toggle visibility of employee form//
     @FXML
     private AnchorPane updateView;
     @FXML
     private AnchorPane addEmployeeView;
 
-    ObservableList<Employee> observableList = FXCollections.observableList(EmployeeDAO.getEmployees());
     private Employee employee;
-    //
+
     @FXML
     private void initialize() {
-        employeeListTableViewFirstName.setCellValueFactory(new PropertyValueFactory<>("firstName"));
-        employeeListTableViewLastName.setCellValueFactory(new PropertyValueFactory<>("lastName"));
-        employeeListTableViewEmail.setCellValueFactory(new PropertyValueFactory<>("email"));
-        employeeListTableViewBirthDate.setCellValueFactory(new PropertyValueFactory<>("birthDate"));
-        employeeListTableViewSalary.setCellValueFactory(new PropertyValueFactory<>("salary"));
-
-        employeeListTableView.setItems(observableList);
+        employeeTableView();
     }
 
     public void saveEmployee() {
-        employee = new Employee(
-                firstName.getText(),
-                lastName.getText(),
-                email.getText(),
-                birthDate.getValue(),
-                Double.parseDouble(salary.getText()));
+        boolean isEmailValid = DataValidation.emailValidation(email, "Please enter a valid email!");
+        boolean isFirstNameValid = DataValidation.textValidation(firstName, "Please enter a valid name!");
+        boolean isLastNameValid = DataValidation.textValidation(lastName, "Please enter a valid name!");
+        boolean isSalaryValid = DataValidation.doubleValidator(salary, "Please enter valid salary!");
 
-        EmployeeDAO.saveEmployee(employee);
-        addEmployeeView.setVisible(false);
+        if (isEmailValid && isFirstNameValid && isLastNameValid && isSalaryValid) {
+            employee = new Employee(
+                    firstName.getText(),
+                    lastName.getText(),
+                    email.getText(),
+                    birthDate.getValue(),
+                    Double.parseDouble(salary.getText()));
+
+            EmployeeDAO.saveEmployee(employee);
+            employeeTableView();
+            addEmployeeView.setVisible(false);
+        }
     }
 
-    //toggle visibility of employee form
     public void toggleAddEmployee() {
         boolean flag = addEmployeeView.isVisible();
         addEmployeeView.setVisible(!flag);
+        updateView.setVisible(false);
     }
 
     public void updateEmployee() {
-        boolean flag = updateView.isVisible();
+        boolean isEmailValid = DataValidation.emailValidation(emailUpd, "Please enter a valid email!");
+        boolean isFirstNameValid = DataValidation.textValidation(firstNameUpd, "Please enter a valid name!");
+        boolean isLastNameValid = DataValidation.textValidation(lastNameUpd, "Please enter a valid name!");
+        boolean isSalaryValid = DataValidation.doubleValidator(salaryUpd, "Please enter valid salary!");
 
-        employee = employeeListTableView.getSelectionModel().getSelectedItem();
-        employee.setFirstName(firstNameUpd.getText());
-        employee.setLastName(lastNameUpd.getText());
-        employee.setEmail(emailUpd.getText());
-        employee.setBirthDate(birthDateUpd.getValue());
-        employee.setSalary(Double.parseDouble(salaryUpd.getText()));
+        if (isEmailValid && isFirstNameValid && isLastNameValid && isSalaryValid) {
+            employee = employeeListTableView.getSelectionModel().getSelectedItem();
 
-        EmployeeDAO.saveOrUpdateEmployee(employee);
-        updateView.setVisible(!flag);
+            employee.setFirstName(firstNameUpd.getText());
+            employee.setLastName(lastNameUpd.getText());
+            employee.setEmail(emailUpd.getText());
+            employee.setBirthDate(birthDateUpd.getValue());
+            employee.setSalary(Double.parseDouble(salaryUpd.getText()));
+
+            EmployeeDAO.saveOrUpdateEmployee(employee);
+            updateView.setVisible(false);
+            employeeTableView();
+        }
     }
 
     public void toggleUpdate() {
         boolean flag = updateView.isVisible();
-        updateView.setVisible(!flag);
         if (employeeListTableView.getSelectionModel().getSelectedItem() != null) {
             employee = employeeListTableView.getSelectionModel().getSelectedItem();
+
             firstNameUpd.setText(employee.getFirstName());
             lastNameUpd.setText(employee.getLastName());
             emailUpd.setText(employee.getEmail());
             birthDateUpd.setValue(LocalDate.parse(employee.getBirthDate().toString()));
             salaryUpd.setText(Double.toString(employee.getSalary()));
+
+            updateView.setVisible(!flag);
+            addEmployeeView.setVisible(false);
         }
     }
 
@@ -130,8 +140,23 @@ public class EmployeeController {
     }
 
     public void deleteEmployee() {
-        employee = employeeListTableView.getSelectionModel().getSelectedItem();
-        EmployeeDAO.deleteEmployee(employee);
+        if (employeeListTableView.getSelectionModel().getSelectedItem() != null) {
+            employee = employeeListTableView.getSelectionModel().getSelectedItem();
+            EmployeeDAO.deleteEmployee(employee);
+            employeeTableView();
+        }
+    }
+
+    private void employeeTableView() {
+        ObservableList<Employee> observableList = FXCollections.observableList(EmployeeDAO.getEmployees());
+
+        employeeListTableViewFirstName.setCellValueFactory(new PropertyValueFactory<>("firstName"));
+        employeeListTableViewLastName.setCellValueFactory(new PropertyValueFactory<>("lastName"));
+        employeeListTableViewEmail.setCellValueFactory(new PropertyValueFactory<>("email"));
+        employeeListTableViewBirthDate.setCellValueFactory(new PropertyValueFactory<>("birthDate"));
+        employeeListTableViewSalary.setCellValueFactory(new PropertyValueFactory<>("salary"));
+
+        employeeListTableView.setItems(observableList);
     }
 
     public void backToMain(ActionEvent actionEvent) throws IOException {
